pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag-server"
version = "0.1.0"
description = "RAG 能力的 API 服务，提供检索增强生成接口"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "CHyfuture"}
]
keywords = ["rag", "api", "retrieval", "llm", "vector"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "milvus-service @ git+ssh://git@github.com/CHyfuture/BaseVector-Core.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
agentic-rag-server = "app.main:run"

[project.urls]
Repository = "https://github.com/CHyfuture/agentic-rag-server"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
